+++ spec/unit/alf-core/reader/test_initialize.rb
@@ -1,29 +1,61 @@
 require 'spec_helper'
 module Alf
   describe "Reader#initialize" do
-    
+
     subject{ Reader.new(*args) }
-    
-    describe "with full args" do
-      let(:args){[
-        "suppliers", Environment.examples, {:opts => true}
-      ]}
-      specify {
-        subject.input.should eq("suppliers")
-        subject.environment.should be_a(Environment)
-        subject.options.should eq(:opts => true)
-      }
+
+    let(:input){ nil }
+    let(:environment){ nil }
+    let(:options){ nil }
+
+    let(:args){ [input, environment, options].compact }
+
+    before do
+      subject.should be_a(Reader)
     end
-    
-    describe "with only options" do
-      let(:args){[
-        {:opts => true}
-      ]}
-      specify {
-        subject.input.should be_nil
-        subject.environment.should be_nil
-        subject.options.should eq(:opts => true)
-      }
+
+    after do
+      subject.input.should eq(input)
+      subject.environment.should eq(environment)
+    end
+
+    context 'with empty default options' do
+
+      after do
+        subject.options.should eq(options || {})
+      end
+
+      context "with only a path as a String" do
+        let(:input){ "suppliers" }
+        it 'should set the path correctly' do
+          subject.path.should eq(Path("suppliers"))
+        end
+      end
+
+      context "with only a path as a Path" do
+        let(:input){ Path("suppliers") }
+        it 'should set the path correctly' do
+          subject.path.should eq(Path("suppliers"))
+        end
+      end
+
+      context "with a File" do
+        let(:input){ Path.here.open('r') }
+        it 'should set the path correctly' do
+          subject.path.should eq(Path.here)
+        end
+        after{ input.close rescue nil }
+      end
+
+      describe "with full args" do
+        let(:input){ "suppliers" }
+        let(:environment){ Environment.examples }
+        let(:options){ {:opts => true} }
+        it 'should set the path correctly' do
+          subject.path.should eq(Path("suppliers"))
+        end
+      end
+
     end
 
     describe "on a subclass" do
@@ -31,30 +63,24 @@ module Alf
         DEFAULT_OPTIONS = {:from => :subclass}
       end
       subject{ FooReader.new(*args) }
-        
+
+      let(:input){ "suppliers" }
+
       describe "without option overriding" do
-        let(:args){[
-          {:opts => true}
-        ]}
-        specify {
-          subject.input.should be_nil
-          subject.environment.should be_nil
+        let(:options){ {:opts => true} }
+        it 'should merge options' do
           subject.options.should eq(:opts => true, :from => :subclass)
-        }
+        end
       end
-      
+
       describe "with option overriding" do
-        let(:args){[
-          {:opts => true, :from => :overrided}
-        ]}
-        specify {
-          subject.input.should be_nil
-          subject.environment.should be_nil
+        let(:options){ {:opts => true, :from => :overrided} }
+        it 'should override options' do
           subject.options.should eq(:opts => true, :from => :overrided)
-        }
+        end
       end
-      
+
     end
-    
+
   end
 end